Use keyname for page streaming fields.
