+++ src/main/java/com/google/api/codegen/transformer/php/PhpSurfaceNamer.java
@@ -17,6 +17,7 @@ package com.google.api.codegen.transformer.php;
 import com.google.api.codegen.ServiceMessages;
 import com.google.api.codegen.config.InterfaceConfig;
 import com.google.api.codegen.config.MethodConfig;
+import com.google.api.codegen.config.PageStreamingConfig;
 import com.google.api.codegen.config.SingleResourceNameConfig;
 import com.google.api.codegen.config.VisibilityConfig;
 import com.google.api.codegen.transformer.ModelTypeFormatterImpl;
@@ -177,5 +178,29 @@ public class PhpSurfaceNamer extends SurfaceNamer {
   @Override
   public boolean methodHasRetrySettings(MethodConfig methodConfig) {
     return !methodConfig.isGrpcStreaming();
+  }
+
+  public String getRequestTokenFieldName(PageStreamingConfig pageStreaming) {
+    // Not using keyName since PHP maps the string "requestPageTokenField" to the raw request token
+    // field name from the page streaming config.
+    return pageStreaming.getRequestTokenField().getSimpleName();
+  }
+
+  public String getPageSizeFieldName(PageStreamingConfig pageStreaming) {
+    // Not using keyName since PHP maps the string "requestPageSizeField" to the raw page size
+    // field name from the page streaming config.
+    return pageStreaming.getPageSizeField().getSimpleName();
+  }
+
+  public String getResponseTokenFieldName(PageStreamingConfig pageStreaming) {
+    // Not using keyName since PHP maps the string "responsePageTokenField" to the raw response
+    // token field name from the page streaming config.
+    return pageStreaming.getResponseTokenField().getSimpleName();
+  }
+
+  public String getResourcesFieldName(PageStreamingConfig pageStreaming) {
+    // Not using keyName since PHP maps the string "resourceField" to the raw resource field
+    // from name the page streaming config.
+    return pageStreaming.getResourcesFieldName();
   }
 }
